fix: check if account active
