+++ app/(tabs)/profile/_components/provider/profile.tsx
@@ -1,5 +1,5 @@
+import { ScrollView, StyleSheet } from 'react-native';
 import { Flex } from 'react-native-flex';
-import { ScrollView } from 'react-native-gesture-handler';
 import OptionsCustomerProfile from '../customer/customer-options';
 import BannerProfile from './banner';
 
@@ -7,7 +7,7 @@ export default function ProfileProvider() {
   return (
     <Flex vertical>
       <BannerProfile />
-      <Flex mt={90}>
+      <Flex mt={100}>
         <ScrollView>
           <OptionsCustomerProfile />
         </ScrollView>
@@ -15,3 +15,17 @@ export default function ProfileProvider() {
     </Flex>
   );
 }
+
+const styles = StyleSheet.create({
+  scene: {
+    flex: 1,
+    alignItems: 'center',
+    marginTop: 20,
+  },
+  p: {
+    fontSize: 14,
+    color: '#555',
+    lineHeight: 22,
+    marginBottom: 14,
+  },
+});
