+++ package.json
@@ -55,11 +55,13 @@
     "react-native-maps": "^1.25.2",
     "react-native-mask-input": "^1.2.3",
     "react-native-mmkv": "^3.3.0",
+    "react-native-pager-view": "6.7.1",
     "react-native-portalize": "^1.0.7",
     "react-native-reanimated": "^4.1.3",
     "react-native-safe-area-context": "5.4.0",
     "react-native-screens": "~4.11.1",
     "react-native-svg": "15.11.2",
+    "react-native-tab-view": "^4.2.0",
     "react-native-vision-camera": "^4.7.2",
     "react-native-web": "~0.20.0",
     "react-native-webview": "13.13.5",
