+++ store/users/useEnableAccount.ts
@@ -1,24 +1,15 @@
 import api from '@/services/api';
 import { useMutation } from '@tanstack/react-query';
+import { UserResponse } from '../session/useSignin';
 
 interface User {
   code: string;
   userId: string;
 }
 
-interface AccountResponse {
-  accessToken: string;
-  refreshToken: string;
-  id: string;
-  name: string;
-  email: string;
-  phone: string;
-  photo: string | null;
-}
-
 export function useEnableAccount() {
   const mutationFn = async (enableAccountData: User) => {
-    const { data } = await api.post<AccountResponse>(
+    const { data } = await api.post<UserResponse>(
       '/session/code-check',
       enableAccountData,
     );
